Extracting common functionality from tests checking for the apps directory into the getAllLayoutSets.ts file
